♻️ refactor(websearch): separate fetch logic into reusable tool
- create new fetch module with content extraction capabilities
- move common web scraping logic to Fetch class
- update websearch tools to use the new fetch implementation
- consolidate user agent generation in utils module
- remove duplicated code across search implementations
